Update app.js
Messy right now. Learned about turf.js, a geospatial analysis js library that can calculate routes given the curvature of the earth for advanced geospatial queries and analysis. So far, incorporated the bbox method from turf to calculate the bounding box for each zone selected and storing the results in the selectedZoneBBOX obj. Figured this is smarter since the key is the zone but this should all be compiled into a single object and updated non-destructively going forward. For now, we have the zone coords and when calling the google places api, I can use a forEach on the bbox obj and store the results in google places api in an object to iterate over along with filtering for only the places that are within the selected with turfs booleanpointinpolygon method. More on that in the next git commit