Update app.js
Completed function queryPlacesAPI(searchTerm). 

usage: click a zone and when highlighted in green, call queryPlacesAPI(keyword) and places in the zone area will populate if the keyword matches place address, description, name, other metadata. I will be storing the places returned in an object called returnedPlacesPerZone where the zone is the key and the results array is the value.

Next is rendering the results on the map with a marker that when clicked, will show a snippet of the place like the business photo, whether it's open or closed based on open times, phone number, website, and directions using the user's geolocation as the origin destination. If geolocation is disabled, the origin address will need to be filled by the user using an autocomplete text while the destination address will be filled by the place direction call.